+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.3.0] - 2025-12-22
+
+_Stable release based on [4.3.0-rc.1]._
+
+## [4.3.0-rc.1] - 2025-12-22
+
+### Changed
+
+- Log warning when notification failed and error when failed permanently.
+
 ## [4.2.8] - 2025-12-21
 
 ### Fixed
@@ -1010,6 +1020,8 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[4.3.0]: https://https://github.com/internetguru/laravel-common/compare/v4.2.8...v4.3.0
+[4.3.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v4.2.8
 [4.2.8]: https://https://github.com/internetguru/laravel-common/compare/v4.2.7...v4.2.8
 [4.2.7]: https://https://github.com/internetguru/laravel-common/compare/v4.2.6...v4.2.7
 [4.2.6]: https://https://github.com/internetguru/laravel-common/compare/v4.2.5...v4.2.6